refactor(tools): changing tools module path
