+++ scannettes/models/product.py
@@ -5,7 +5,7 @@ from dataclasses import dataclass, field
 from typing import Any, Dict, List, Optional, Union
 
 from scannettes.models.state_handler import State, ProductState
-from scannettes.utils import generate_uuid, update_object
+from scannettes.tools.utils import generate_uuid, update_object
 
 Payload = Dict[str, Any]
