+++ scannettes/models/room.py
@@ -8,7 +8,7 @@ from typing import Any, Dict, Literal, Optional, Union
 from scannettes.models.purchase import Inventory, Purchase
 from scannettes.models.state_handler import RoomState, State
 from scannettes.odoo.odoo import Odoo
-from scannettes.utils import generate_uuid, update_object, restrfmtdate
+from scannettes.tools.utils import generate_uuid, update_object, restrfmtdate
 
 Payload = Dict[str, Any]
 RoomType = Literal["purchase", "inventory"]
